updated check distance and turnOnThermometer function
Added trigPin and echoPin as arguments to check distance function. Added therm object as argument to turnOnThermometer function.

+++ main/contactless_therm.h
@@ -8,7 +8,7 @@
 #include <SparkFunMLX90614.h>
 
 boolean checkDistance();
-int turnOnThermometer();
+float turnOnThermometer();
 void displayTemp(float number);
